+++ packages/analytics_hub_firebase/lib/src/provider/firebase_analytics_hub_provider_identifier.dart
@@ -1,14 +1,11 @@
 import 'package:analytics_hub/analytics_hub.dart';
 import 'package:firebase_core/firebase_core.dart';
 
-import '../resolver/firebase_analytics_event_resolver.dart';
-
 /// [ProviderIdentifier] for the Firebase Analytics provider.
 ///
 /// Use this identifier in [Event.providers] to route events to
 /// [FirebaseAnalyticsHubProvider]. [name] defaults to [defaultFirebaseAppName].
-class FirebaseAnalyticsHubProviderIdentifier
-    extends ProviderIdentifier<FirebaseAnalyticsEventResolver> {
+class FirebaseAnalyticsHubProviderIdentifier extends ProviderIdentifier {
   /// Creates an identifier with optional [name]; defaults to [defaultFirebaseAppName].
   const FirebaseAnalyticsHubProviderIdentifier({
     super.name = defaultFirebaseAppName,