+++ packages/analytics_hub_firebase/lib/src/resolver/firebase_analytics_event_resolver.dart
@@ -1,25 +1,18 @@
 import 'package:analytics_hub/analytics_hub.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 
-import 'firebase_analytics_e_commerce_event_resolver.dart';
-
-/// Resolver that forwards [LogEvent] and [ECommerceEvent] to Firebase Analytics.
+/// Resolver that forwards [Event] to Firebase Analytics.
 ///
-/// Used internally by [FirebaseAnalyticsHubProvider]. [LogEvent] is sent via
-/// [FirebaseAnalytics.logEvent]; e-commerce events are delegated to
-/// [FirebaseAnalyticsECommerceEventResolver].
-class FirebaseAnalyticsEventResolver
-    implements EventResolver, LogEventResolver, ECommerceEventResolver {
+/// Used internally by [FirebaseAnalyticsHubProvider]. [Event] is sent via
+/// [FirebaseAnalytics.logEvent].
+class FirebaseAnalyticsEventResolver implements EventResolver {
   /// Creates a resolver that uses the given [FirebaseAnalytics] instance.
-  FirebaseAnalyticsEventResolver(this._analytics)
-      : _eCommerceEventResolver =
-            FirebaseAnalyticsECommerceEventResolver(_analytics);
+  FirebaseAnalyticsEventResolver(this._analytics);
 
   final FirebaseAnalytics _analytics;
-  final FirebaseAnalyticsECommerceEventResolver _eCommerceEventResolver;
 
   @override
-  Future<void> resolveLogEvent(LogEvent event) {
+  Future<void> resolveEvent(Event event) {
     final parameters = switch (event.properties) {
       null => null,
       final properties => Map<String, Object>.fromEntries(
@@ -34,8 +27,4 @@ class FirebaseAnalyticsEventResolver
       parameters: parameters,
     );
   }
-
-  @override
-  Future<void> resolveECommerceEvent(ECommerceEvent event) =>
-      _eCommerceEventResolver.resolve(event);
 }