+++ packages/analytics_hub_mixpanel/lib/analytics_hub_mixpanel.dart
@@ -1,7 +1,8 @@
 /// Mixpanel implementation for [AnalyticsHub].
 ///
 /// This library provides [MixpanelAnalyticsHubProvider] and
-/// [MixpanelAnalyticsHubProviderKey] to send [LogEvent]s from [AnalyticsHub]
+/// [MixpanelAnalyticsHubProviderIdentifier] to send [Event]s from
+/// [AnalyticsHub]
 /// to Mixpanel via [Mixpanel.track]. Session updates set the Mixpanel identity
 /// via [Mixpanel.identify] or [Mixpanel.reset].
 ///